feat: recommendation

+++ internal/app/product_recommendation/service/product_recommendation_delete_service.go
@@ -1,6 +1,13 @@
 package service
 
+import "fmt"
+
 func (pr *ProductRecommendationServiceImpl) Delete(id int) error {
+
+	ifExist, _ := pr.ProductRecommendationRepository.GetById(id)
+	if ifExist == nil {
+		return fmt.Errorf("product recommendation not found")
+	}
 
 	err := pr.ProductRecommendationRepository.Delete(id)
 	if err != nil {
